docker/pyproject.toml
[project]
name = "trent_docker"
version = "2.0.0"
description = "Unified MCP Server: RAG, Research, MediaWiki, and Template Installation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "trent" }
]

dependencies = [
    # Core MCP
    "fastmcp>=2.2.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.30.0",

    # RAG / Database
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",

    # SQL Parsing
    "sqlparse>=0.5.3",

    # Research Tools (NO DuckDuckGo!)
    "anthropic>=0.66.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.12.15",
    "requests>=2.32.5",
    "pymupdf>=1.26.4",
    "pydantic>=2.0.0",

    # MediaWiki
    # Note: No atlassian-python-api - MediaWiki uses requests

    # Template Installer utilities
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
trent = "trent.server:main"
trent_http = "trent.server:main_http"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trent"]
